split of hardware handling from HAL_Device.py to new HAL_ShimMQ.py
